+++ bootstrap/http/start.php
@@ -2,7 +2,8 @@
 /**
  * Boots up our application with an HTTP kernel
  */
-use RDev\Applications\Bootstrappers\IO\BootstrapperIO;
+use RDev\Applications\Bootstrappers\ApplicationBinder;
+use RDev\Applications\Bootstrappers\IO\IBootstrapperIO;
 use RDev\Framework\HTTP\Kernel;
 use RDev\HTTP\Requests\Request;
 use RDev\Routing\Router;
@@ -11,18 +12,16 @@ require_once __DIR__ . "/../start.php";
 
 /**
  * ----------------------------------------------------------
- * Setup the bootstrappers
+ * Finish configuring the bootstrappers for the HTTP kernel
  * ----------------------------------------------------------
  *
- * @var Closure $configureBootstrappers
+ * @var ApplicationBinder $applicationBinder
  */
-$configureBootstrappers = require __DIR__ . "/../configureBootstrappers.php";
-$configureBootstrappers(
-    $application,
-    require $application->getPaths()["configs"] . "/http/bootstrappers.php",
+$applicationBinder->bindToApplication(
+    require __DIR__ . "/../../configs/http/bootstrappers.php",
     false,
     true,
-    $application->getPaths()["tmp.framework.http"] . "/" . BootstrapperIO::DEFAULT_CACHED_REGISTRY_FILE_NAME
+    $application->getPaths()["tmp.framework.http"] . "/" . IBootstrapperIO::DEFAULT_CACHED_REGISTRY_FILE_NAME
 );
 
 /**
@@ -40,10 +39,11 @@ $application->start(function() use ($application)
      * @var Router $router
      * @var Request $request
      */
-    $router = $application->getIoCContainer()->makeShared("RDev\\Routing\\Router");
-    $request = $application->getIoCContainer()->makeShared("RDev\\HTTP\\Requests\\Request");
-    $kernel = new Kernel($application->getIoCContainer(), $router, $application->getLogger());
-    $kernel->addMiddleware(require $application->getPaths()["configs"] . "/http/middleware.php");
+    $router = $application->getIoCContainer()->makeShared(Router::class);
+    $request = $application->getIoCContainer()->makeShared(Request::class);
+    $logger = require __DIR__ . "/../../configs/http/logging.php";
+    $kernel = new Kernel($application->getIoCContainer(), $router, $logger);
+    $kernel->addMiddleware(require __DIR__ . "/../../configs/http/middleware.php");
     $response = $kernel->handle($request);
     $response->send();
 });