+++ configs/tasks.php
@@ -1,0 +1,9 @@
+<?php
+/**
+ * Copyright (C) 2015 David Young
+ * 
+ * Defines the application task config
+ */
+use RDev\Applications\Tasks\Dispatchers\Dispatcher;
+
+return new Dispatcher();